+++ astronomy/orbit_analysis_test.cpp
@@ -13,9 +13,9 @@
 #include "physics/ephemeris.hpp"
 #include "physics/solar_system.hpp"
 #include "testing_utilities/approximate_quantity.hpp"
-#include "testing_utilities/matchers.hpp"
-#include "testing_utilities/numerics.hpp"
+#include "testing_utilities/error_matchers.hpp"
 #include "testing_utilities/is_near.hpp"
+#include "testing_utilities/matchers.hpp"
 
 namespace principia {
 namespace astronomy {
