+++ astronomy/orbital_elements_test.cpp
@@ -16,9 +16,9 @@
 #include "physics/solar_system.hpp"
 #include "testing_utilities/almost_equals.hpp"
 #include "testing_utilities/approximate_quantity.hpp"
+#include "testing_utilities/error_matchers.hpp"
 #include "testing_utilities/is_near.hpp"
 #include "testing_utilities/matchers.hpp"
-#include "testing_utilities/numerics.hpp"
 
 namespace principia {
