+++ testing_utilities/numerics.hpp
@@ -6,17 +6,10 @@
 #include "geometry/grassmann.hpp"
 #include "geometry/point.hpp"
 #include "geometry/r3_element.hpp"
-#include "gmock/gmock.h"
 #include "quantities/quantities.hpp"
 
 namespace principia {
 namespace testing_utilities {
-namespace internal_numerics {
-
-using quantities::Difference;
-using ::testing::Matcher;
-using ::testing::MatcherInterface;
-using ::testing::MatchResultListener;
 
 template<typename Scalar>
 double DoubleValue(Scalar const& scalar);
@@ -85,79 +78,6 @@ double RelativeError(geometry::R3Element<Scalar> const& expected,
 template<typename Scalar, typename Frame, int rank>
 double RelativeError(geometry::Multivector<Scalar, Frame, rank> const& expected,
                      geometry::Multivector<Scalar, Frame, rank> const& actual);
-
-template<typename Value>
-class DifferenceFromMatcher : public MatcherInterface<Value const&> {
- public:
-  DifferenceFromMatcher(Value const& expected,
-                        Matcher<Difference<Value>> const& error_matcher);
-
-  bool MatchAndExplain(Value const& actual,
-                       MatchResultListener* listener) const override;
-  void DescribeTo(std::ostream* os) const override;
-  void DescribeNegationTo(std::ostream* os) const override;
-
- private:
-  Value expected_;
-  Matcher<Difference<Value>> error_matcher_;
-};
-
-template<typename Value>
-class AbsoluteErrorFromMatcher : public MatcherInterface<Value const&> {
- public:
-  using Error =
-      decltype(AbsoluteError(std::declval<Value>(), std::declval<Value>()));
-
-  AbsoluteErrorFromMatcher(Value const& expected,
-                           Matcher<Error> const& error_matcher);
-
-  bool MatchAndExplain(Value const& actual,
-                       MatchResultListener* listener) const override;
-  void DescribeTo(std::ostream* os) const override;
-  void DescribeNegationTo(std::ostream* os) const override;
-
- private:
-  Value expected_;
-  Matcher<Error> error_matcher_;
-};
-
-template<typename Value>
-class RelativeErrorFromMatcher : public MatcherInterface<Value const&> {
- public:
-  RelativeErrorFromMatcher(Value const& expected,
-                           Matcher<double> const& error_matcher);
-
-  bool MatchAndExplain(Value const& actual,
-                       MatchResultListener* listener) const override;
-  void DescribeTo(std::ostream* os) const override;
-  void DescribeNegationTo(std::ostream* os) const override;
-
- private:
-  Value expected_;
-  Matcher<double> error_matcher_;
-};
-
-template<typename Value>
-Matcher<Value const&> DifferenceFrom(
-    Value const& expected,
-    Matcher<Difference<Value>> const& error_matcher);
-
-template<typename Value, typename ErrorMatcher>
-Matcher<Value const&> AbsoluteErrorFrom(Value const& expected,
-                                        ErrorMatcher const& error_matcher);
-
-template<typename Value>
-Matcher<Value const&> RelativeErrorFrom(Value const& expected,
-                                        Matcher<double> const& error_matcher);
-
-}  // namespace internal_numerics
-
-using internal_numerics::AbsoluteError;
-using internal_numerics::AbsoluteErrorFrom;
-using internal_numerics::DifferenceFrom;
-using internal_numerics::DoubleValue;
-using internal_numerics::RelativeError;
-using internal_numerics::RelativeErrorFrom;
 
 }  // namespace testing_utilities
 }  // namespace principia
