+++ testing_utilities/testing_utilities.vcxproj.filters
@@ -92,6 +92,12 @@
     <ClInclude Include="approximate_quantity_body.hpp">
       <Filter>Source Files</Filter>
     </ClInclude>
+    <ClInclude Include="error_matchers.hpp">
+      <Filter>Header Files</Filter>
+    </ClInclude>
+    <ClInclude Include="error_matchers_body.hpp">
+      <Filter>Source Files</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="almost_equals_test.cpp">
